requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
html5lib==1.1
httpcache==0.1.3
idna==2.10
Pillow==7.2.0
pyaml==20.4.0
PyYAML==5.3.1
requests==2.24.0
singleton-factory==1.0
singleton3==1.0
six==1.3.0
soupsieve==2.0.1
urllib3==1.25.9
webencodings==0.5.1
